+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 - Implementado NT2024.003
 - Adicionado Evento Insucesso Entrega (NT2023.005)
 - Adicionado Evento Cancelamento Insucesso Entrega (NT2023.005)
+- Adicionado Evento Conciliacao Financeira (NT2024.002)
+- Adicionado Evento Cancelamento Conciliacao Financeira (NT2024.002)